+++ src/data/arbitrary/mod.rs
@@ -8,7 +8,8 @@
 //!
 //! **Note**
 //!
-//! You must enable the *arbitrary* feature in your `Cargo.toml` to use these functions.
+//! You must enable the *fuzzing* feature in your `Cargo.toml` to use these
+//! functions.
 //!
 //! </div>